invalidate leaflet map size when toggleFeatureLayer
