+++ frontend/src/stores/map.js
@@ -1,6 +1,9 @@
 import { defineStore } from 'pinia'
-import { ref, shallowRef } from 'vue'
-import L from 'leaflet'
+import { nextTick, ref, shallowRef } from 'vue'
+import L, { canvas } from 'leaflet'
+import * as esriLeaflet from 'esri-leaflet'
+import * as esriLeafletGeocoder from 'esri-leaflet-geocoder'
+import { useFeaturesStore } from '@/stores/features'
 
 export const useMapStore = defineStore('map', () => {
   const leaflet = shallowRef(null)
@@ -20,6 +23,7 @@ export const useMapStore = defineStore('map', () => {
   const mapLoaded = ref(false)
 
   const MIN_WMS_ZOOM = 9
+  const MIN_WFS_ZOOM = 9
   const COMRES_SERVICE_URL = 'https://arcgis.cuahsi.org/arcgis/services/CIROH-ComRes'
   const COMRES_REST_URL = 'https://arcgis.cuahsi.org/arcgis/rest/services/CIROH-ComRes'
 
@@ -57,16 +61,21 @@ export const useMapStore = defineStore('map', () => {
 
   const limitToBounds = (featureLayer) => {
     console.log('Limiting to bounds of feature layer:', featureLayer)
-    featureLayer.query().bounds(function (error, bounds) {
+    featureLayer.query().bounds(async function (error, bounds) {
       if (error) {
         console.log('Error running bounds query:')
         console.warn(error)
       }
       try {
-        console.log(`Zooming to bounds: ${bounds}`)
-        leaflet.value.fitBounds(bounds)
+        console.log('Setting bounds to:', bounds)
+        leaflet.value.setMaxBounds(null)
+        leaflet.value.setView([0, 0], 2)
+        leaflet.value.invalidateSize()
+
         // prevent panning from bounds
         leaflet.value.setMaxBounds(bounds)
+        leaflet.value.fitBounds(bounds)
+        await nextTick()
         // prevent zooming out
         leaflet.value.setMinZoom(leaflet.value.getZoom())
       } catch (error) {
@@ -90,9 +99,8 @@ export const useMapStore = defineStore('map', () => {
     const response = await fetch(queryUrl)
     const data = await response.json()
     if (data && data.layers) {
-      console.log(`Creating WMS Layers for ${region.name}:`)
+      console.log(`Creating WMS Layers for ${region.name}`)
       data.layers.forEach((layer) => {
-        console.log(`Creating layer ${layer.id}: ${layer.name}...`)
         url = `${COMRES_SERVICE_URL}/${region.name}/MapServer/WmsServer?`
         // https://leafletjs.com/reference.html#tilelayer-wms
         const wmsLayer = L.tileLayer.wms(url, {
@@ -113,21 +121,18 @@ export const useMapStore = defineStore('map', () => {
         wmsLayers.value[region.name] = wmsLayers.value[region.name] || []
         wmsLayers.value[region.name].push(wmsLayer)
       })
-      console.log('WMS Layers thus far:', wmsLayers.value)
     } else {
       console.error(`No layers found for ${region.name}`)
     }
   }
 
   async function addWMSLayers(region) {
-    // await createWMSLayers(region)
     for (let layer of wmsLayers.value[region.name] || []) {
       // only add layers that are part of the current region
       if (!layer.name.includes(region.name)) {
-        console.log(`Skipping layer ${layer.name} for region ${region.name}`)
+        console.warn(`Skipping layer ${layer.name} for region ${region.name}`)
         continue
       }
-      console.log('Adding WMS layer:', layer.name)
       layer.addTo(leaflet.value)
       control.value.addOverlay(layer, layer.name)
     }
@@ -136,10 +141,8 @@ export const useMapStore = defineStore('map', () => {
   const toggleWMSLayers = (region) => {
     // if the regionName is not in wmsLayers, create it
     if (!wmsLayers.value[region.name]) {
-      console.log(`Creating WMS layers for region: ${region.name}`)
       createWMSLayers(region)
         .then(() => {
-          console.log(`WMS layers created for region: ${region.name}`)
           addWMSLayers(region)
         })
         .catch((error) => {
@@ -158,11 +161,89 @@ export const useMapStore = defineStore('map', () => {
     })
   }
 
-  const toggleFeatureLayer = (layerName) => {
+  function createFlowlinesFeatureLayer(region) {
+    const featureStore = useFeaturesStore()
+    let url = `https://arcgis.cuahsi.org/arcgis/rest/services/CIROH-ComRes/${region.name}/FeatureServer/${region.flowlinesLayerNumber}`
+    const featureLayer = esriLeaflet.featureLayer({
+      url: url,
+      simplifyFactor: 0.4,
+      precision: 5,
+      minZoom: MIN_WFS_ZOOM,
+      renderer: canvas({ tolerance: 5 }),
+      color: featureOptions.value.defaultColor,
+      weight: featureOptions.value.defaultWeight,
+      opacity: featureOptions.value.opacity,
+      fields: [
+        'FID',
+        'COMID',
+        'REACHCODE',
+        'PopupTitle',
+        'PopupSubti',
+        'SLOPE',
+        'LENGTHKM',
+        'Hydroseq',
+        'GNIS_ID'
+      ]
+    })
+    featureLayer.name = region.name
+
+    featureLayer.on('click', function (e) {
+      const feature = e.layer.feature
+      const properties = feature.properties
+      console.log('Feature clicked:', feature)
+      featureStore.clearSelectedFeatures()
+      if (!featureStore.checkFeatureSelected(feature)) {
+        // Only allow one feature to be selected at a time
+        featureStore.selectFeature(feature)
+      }
+      const popup = L.popup()
+      const content = `
+        ${properties.PopupTitle ? `<h3>${properties.PopupTitle}</h3>` : ''}
+        ${properties.PopupSubti ? `<h4>${properties.PopupSubti}</h4>` : ''}
+        <p>
+            <ul>
+          ${properties.REACHCODE ? `<li>Reach Code: ${properties.REACHCODE}</li>` : ''}
+          ${properties.COMID ? `<li>COMID: ${properties.COMID}</li>` : ''}
+          ${properties.Hydroseq ? `<li>Hydroseq: ${properties.Hydroseq}</li>` : ''}
+          ${properties.SLOPE ? `<li>Slope: ${properties.SLOPE.toFixed(4)}</li>` : ''}
+          ${properties.LENGTHKM ? `<li>Length: ${properties.LENGTHKM.toFixed(4)} km</li>` : ''}
+          ${properties.GNIS_ID ? `<li>GNIS ID: ${properties.GNIS_ID}</li>` : ''}
+            </ul>
+        </p>
+        `
+      popup.setLatLng(e.latlng).setContent(content).openOn(leaflet.value)
+    })
+    return featureLayer
+  }
+
+  function createFeatureLayerProvider(region) {
+    let url = `https://arcgis.cuahsi.org/arcgis/rest/services/CIROH-ComRes/${region.name}/FeatureServer/${region.flowlinesLayerNumber}`
+    const featureLayerProvider = esriLeafletGeocoder.featureLayerProvider({
+      url: url,
+      searchFields: ['PopupTitle'],
+      label: `${region.name} Flowlines`,
+      //bufferRadius: 5000,
+      formatSuggestion: function (feature) {
+        return feature.properties.PopupTitle
+      }
+    })
+    return featureLayerProvider
+  }
+
+  const toggleFeatureLayer = async (region) => {
+    // first check if the region is already added
+    if (!flowlinesFeatureLayers.value.some((layer) => layer.name === region.name)) {
+      console.log(`Adding feature layer for region: ${region.name}`)
+      const flowlines = createFlowlinesFeatureLayer(region)
+      flowlinesFeatureLayers.value.push(flowlines)
+      const provider = createFeatureLayerProvider(region)
+      featureLayerProviders.value.push(provider)
+      region.flowlinesLayer = flowlines
+    }
     // turn off all other feature layers and turn on the selected one
-    console.log('Toggling feature layer:', layerName)
+    console.log('Toggling feature layer:', region.name)
     flowlinesFeatureLayers.value.forEach((featureLayer) => {
-      if (featureLayer.name !== layerName) {
+      if (featureLayer.name !== region.name) {
         featureLayer.removeFrom(leaflet.value)
         control.value.removeLayer(featureLayer)
       } else {
